feat: Implement product management interface with CRUD functionality
- Added product table view with search and pagination features.
- Implemented create, edit, and delete functionalities for products.
- Enhanced user experience with confirmation dialogs for delete actions.
- Integrated toast notifications for successful actions in size and store forms.
- Developed a responsive shop page with sorting and filtering options.
- Created a quick view modal for product details with image selection.
- Added toast component for displaying messages to users.

+++ app/Http/Controllers/BranchController.php
@@ -105,6 +105,11 @@ class BranchController extends Controller
      */
     public function destroy(string $id)
     {
+        $record = Branch::find($id);
+
+        if($record->products()->exists()){
+            return redirect()->back()->with("error","Cannot delete this record for the relations !");
+        }
         $deleted = $this->crud->delete($id);
         if($deleted){
             return redirect()->back()->with("success","Delete Record Success !");
