+++ app/Http/Controllers/ColorController.php
@@ -88,6 +88,11 @@ class ColorController extends Controller
      */
     public function destroy(string $id)
     {
+        $record = Color::find($id);
+
+        if($record->products()->exists()){
+            return redirect()->back()->with("error","Cannot delete this record for the relations !");
+        }
         $deleted = $this->crud->delete($id);
         if($deleted){
             return redirect()->back()->with("success","Delete Record Success !");
