+++ app/Http/Controllers/StoreSectionController.php
@@ -90,6 +90,11 @@ class StoreSectionController extends Controller
      */
     public function destroy(string $id)
     {
+        $record = StoreSections::find($id);
+
+        if($record->products()->exists() || $record->branches()->exists()){
+            return redirect()->back()->with("error","Cannot delete this record for the relations !");
+        }
         $deleted = $this->crud->delete($id);
         if($deleted){
             return redirect()->back()->with("success","Delete Record Success !");
